+++ sdk/dotnet/Compute/ResourcePolicy.cs
@@ -36,6 +36,9 @@ namespace Pulumi.Gcp.Compute
         [Output("region")]
         public Output<string> Region { get; private set; } = null!;
 
+        /// <summary>
+        /// The URI of the created resource.
+        /// </summary>
         [Output("selfLink")]
         public Output<string> SelfLink { get; private set; } = null!;
 
@@ -150,6 +153,9 @@ namespace Pulumi.Gcp.Compute
         [Input("region")]
         public Input<string>? Region { get; set; }
 
+        /// <summary>
+        /// The URI of the created resource.
+        /// </summary>
         [Input("selfLink")]
         public Input<string>? SelfLink { get; set; }