+++ sdk/go/gcp/compute/resourcePolicy.go
@@ -24,6 +24,7 @@ type ResourcePolicy struct {
 	Project pulumi.StringOutput `pulumi:"project"`
 	// Region where resource policy resides.
 	Region pulumi.StringOutput `pulumi:"region"`
+	// The URI of the created resource.
 	SelfLink pulumi.StringOutput `pulumi:"selfLink"`
 	// Policy for creating snapshots of persistent disks.
 	SnapshotSchedulePolicy ResourcePolicySnapshotSchedulePolicyPtrOutput `pulumi:"snapshotSchedulePolicy"`
@@ -67,6 +68,7 @@ type resourcePolicyState struct {
 	Project *string `pulumi:"project"`
 	// Region where resource policy resides.
 	Region *string `pulumi:"region"`
+	// The URI of the created resource.
 	SelfLink *string `pulumi:"selfLink"`
 	// Policy for creating snapshots of persistent disks.
 	SnapshotSchedulePolicy *ResourcePolicySnapshotSchedulePolicy `pulumi:"snapshotSchedulePolicy"`
@@ -83,6 +85,7 @@ type ResourcePolicyState struct {
 	Project pulumi.StringPtrInput
 	// Region where resource policy resides.
 	Region pulumi.StringPtrInput
+	// The URI of the created resource.
 	SelfLink pulumi.StringPtrInput
 	// Policy for creating snapshots of persistent disks.
 	SnapshotSchedulePolicy ResourcePolicySnapshotSchedulePolicyPtrInput