+++ sdk/nodejs/compute/resourcePolicy.ts
@@ -52,6 +52,9 @@ export class ResourcePolicy extends pulumi.CustomResource {
      * Region where resource policy resides.
      */
     public readonly region!: pulumi.Output<string>;
+    /**
+     * The URI of the created resource.
+     */
     public /*out*/ readonly selfLink!: pulumi.Output<string>;
     /**
      * Policy for creating snapshots of persistent disks.
@@ -114,6 +117,9 @@ export interface ResourcePolicyState {
      * Region where resource policy resides.
      */
     readonly region?: pulumi.Input<string>;
+    /**
+     * The URI of the created resource.
+     */
     readonly selfLink?: pulumi.Input<string>;
     /**
      * Policy for creating snapshots of persistent disks.